# openapi.yaml
openapi: 3.0.3
info:
  title: Magma API
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/billing/simulations/users/{id}:
    get:
      tags:
        - Billing
      description: Simulates an invoice for a specific user based on current
        consumption and settings.
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  username:
                    type: string
                  period:
                    type: object
                    properties:
                      from:
                        type: string
                      to:
                        type: string
                  lines:
                    type: array
                    items:
                      type: object
                      properties:
                        concept:
                          type: string
                        quantity:
                          type: integer
                        unitPrice:
                          type: number
                        total:
                          type: number
                  total:
                    type: number
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/billing/sync:
    get:
      tags:
        - Billing
      description: List all synchronization events with pagination and filtering.
        Admin only.
      parameters:
        - schema:
            type: integer
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: integer
            default: 20
          in: query
          name: limit
          required: false
        - schema:
            anyOf:
              - type: string
              - type: array
                items:
                  type: string
          in: query
          name: userIds
          required: false
        - schema:
            anyOf:
              - type: string
              - type: array
                items:
                  type: string
          in: query
          name: months
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        userId:
                          type: string
                        username:
                          type: string
                        month:
                          type: string
                        type:
                          type: string
                        quantity:
                          type: integer
                        nexudusSaleId:
                          type: string
                        nexudusCoworkerId:
                          type:
                            - "null"
                            - string
                        createdOn:
                          type: string
                  pagination:
                    type: object
                    properties:
                      total_records:
                        type: integer
                      current_page:
                        type: integer
                      total_pages:
                        type: integer
                      limit:
                        type: integer
    post:
      tags:
        - Billing
      description: Synchronizes current month consumption for a user to Nexudus.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
              required:
                - userId
      responses:
        "200":
          description: Default Response
  /api/v1/billing/sync/users/{id}/status:
    get:
      tags:
        - Billing
      description: Checks if consumption has been synced for a user in the current month.
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /api/v1/billing/sync/{id}:
    get:
      tags:
        - Billing
      description: Returns detail of a specific sync event.
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    delete:
      tags:
        - Billing
      description: Rollback a specific synchronization event. Deletes in Nexudus and
        releases copies in Magma.
      parameters:
        - schema:
            type: boolean
            default: false
          in: query
          name: force
          required: false
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  localId:
                    type: string
                  remoteId:
                    type: integer
                  status:
                    type: string
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/billing/sync/nexudus:
    post:
      tags:
        - Billing
      description: Synchronizes monthly consumption with Nexudus. Admin only.
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  month:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: string
                        username:
                          type: string
                        salesCreated:
                          type: integer
                        skipped:
                          type: integer
                        errors:
                          type: integer
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/billing/stats:
    get:
      tags:
        - Billing
      description: Get global sales and sync statistics. Admin only.
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalSalesThisMonth:
                    type: integer
                  usersPendingSync:
                    type: integer
                  period:
                    type: object
                    properties:
                      from:
                        type: string
                      to:
                        type: string
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...